@@ -24,7 +24,7 @@ public class RobotMob extends AbstractMob {
     public RobotMob(float xPosMeter, float yPosMeter) {
         super("robotMob", xPosMeter, yPosMeter, 5);
         animationManager = new AnimationManager(sprite);
-        animationManager.registerAnimation("robotMob", "walk" , AbstractEntity.ENTITY_TEXTURE_ATLAS, 1 / 16f);
+        animationManager.registerAnimation("robotMob", "walk", AbstractEntity.ENTITY_TEXTURE_ATLAS, 1 / 16f);
     }
 
     @Override
